app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.4.0
pydantic==2.6.3
python-multipart==0.0.9
requests==2.32.3
streamlit==1.31.0

